Refactor API to its own directory module
