+++ src/popup/warnings/edit-page.tsx
@@ -2,7 +2,7 @@ import Layout from '../layout'
 import { useWarning, useUpdateWarningMutation } from './api-hooks'
 import WarningForm from './form'
 import { useParams, useNavigate } from 'react-router'
-import { type Warning } from '../../warnings'
+import { type Warning } from '../../api'
 
 export default function EditWarningPage(): JSX.Element {
   const updateWarningMutation = useUpdateWarningMutation()
