+++ src/popup/warnings/form.tsx
@@ -4,11 +4,7 @@ import { useId } from 'react'
 import { type SubmitHandler, Controller, useForm } from 'react-hook-form'
 import ColorField from '../color-field'
 import { warningStyles } from './friendly-names'
-import {
-  type BannerWarning,
-  type BorderWarning,
-  type Warning,
-} from '../../warnings'
+import { type BannerWarning, type BorderWarning, type Warning } from '../../api'
 
 export interface WarningFormProps {
   onSave?: (warning: Warning) => void
